refactor: Application Service와 Helper를 분리
