+++ src/main/java/roomescape/business/application_service/reader/UserReader.java
@@ -1,4 +1,4 @@
-package roomescape.business.reader;
+package roomescape.business.application_service.reader;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
