+++ src/main/java/roomescape/business/application_service/service/ReservationTimeService.java
@@ -1,4 +1,4 @@
-package roomescape.business.service;
+package roomescape.business.application_service.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
