+++ src/main/java/roomescape/presentation/api/ReservationApiController.java
@@ -13,11 +13,11 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthRequired;
 import roomescape.auth.LoginInfo;
 import roomescape.auth.Role;
+import roomescape.business.application_service.reader.ReservationReader;
+import roomescape.business.application_service.service.ReservationService;
 import roomescape.business.dto.MyReservationDto;
 import roomescape.business.dto.ReservationDto;
 import roomescape.business.model.vo.UserRole;
-import roomescape.business.reader.ReservationReader;
-import roomescape.business.service.ReservationService;
 import roomescape.presentation.dto.request.AdminReservationRequest;
 import roomescape.presentation.dto.request.ReservationRequest;
 import roomescape.presentation.dto.response.MyReservationResponse;
