+++ src/main/java/roomescape/presentation/api/ReservationTimeApiController.java
@@ -12,11 +12,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthRequired;
 import roomescape.auth.Role;
+import roomescape.business.application_service.reader.ReservationTimeReader;
+import roomescape.business.application_service.service.ReservationTimeService;
 import roomescape.business.dto.ReservableReservationTimeDto;
 import roomescape.business.dto.ReservationTimeDto;
 import roomescape.business.model.vo.UserRole;
-import roomescape.business.reader.ReservationTimeReader;
-import roomescape.business.service.ReservationTimeService;
 import roomescape.presentation.dto.request.ReservationTimeRequest;
 import roomescape.presentation.dto.response.BookedReservationTimeResponse;
 import roomescape.presentation.dto.response.ReservationTimeResponse;
