+++ src/test/java/roomescape/business/service/ThemeServiceTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import roomescape.business.application_service.service.ThemeService;
 import roomescape.business.dto.ThemeDto;
 import roomescape.business.model.entity.Theme;
 import roomescape.business.model.repository.Reservations;
